<R11668> Minor changes needed for building with more recent versions of Visual Studio (e.g., VS2010)

+++ Server/CdrTestClient.cpp
@@ -166,7 +166,7 @@ std::string readFile(std::istream& is)
     char buf[4096];
     while (is) {
         is.read(buf, sizeof buf);
-        int n = is.gcount();
+        int n = (int)is.gcount();
         if (n > 0) {
             char* tmp = new char[nBytes + n];
             if (nBytes > 0)
